% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\docType{data}
\name{d}
\alias{d}
\title{Un data frame completamente inventato, con 100 osservazioni (righe)
e 3 colonne.}
\format{
un data frame con 100 righe e 3 colonne:
\describe{
  \item{w}{peso di campionamento (un numero positivo)}
  \item{s}{strato}
  \item{xcat}{una variabile categoriale (3 categorie)}
}
}
\usage{
d
}
\description{
Un data frame completamente inventato, con 100 osservazioni (righe)
e 3 colonne.
}
\keyword{datasets}
