% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{n_strata}
\alias{n_strata}
\title{Estimate of number of observations in strata of data}
\usage{
n_strata(data, weights, strata)
}
\arguments{
\item{data}{a data frame}

\item{weights}{character string of the name of the weights column}

\item{strata}{character string of the name of the strata column}
}
\description{
Estimate of number of observations in strata of data
}
\examples{
str(d)
n_strata(d, weights = "w", strata = "s")

}
