% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhat.R
\name{vhat}
\alias{vhat}
\title{Estimated variance of a total estimation, with strata}
\usage{
vhat(data, INDEX, weights, strata = NULL)
}
\arguments{
\item{data}{a data frame of individual data. In this version, data should be
the complete dataset of the sample.}

\item{INDEX}{a right hand formula whose result is a logical vector that 
indicates the observations for which the variance has to be estimated 
(see the examples below)}

\item{weights}{character string with the name of the column of weights}

\item{strata}{character string with the name of the column that contains the strata}
}
\description{
Estimated variance of a total estimation, with strata
}
\examples{
str(d)

# Without strata (strata = NULL by default)
vhat(d, INDEX = ~xcat == "cat1", weights = "w")

# With strata (column "s")
vhat(d, INDEX = ~xcat == "cat1", weights = "w", strata = "s")

}
