% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{vhat3}
\alias{vhat3}
\title{Estimated variance of a total estimation, with strata (vector version)}
\usage{
vhat3(x, mh, Nh, strata)
}
\arguments{
\item{x}{un vettore numerico con gli elementi di uno strato (di solito i 
pesi di campionamento)}

\item{mh}{vettore con il numero di osservazioni dello strato (può essere
un vettore di lunghezza uno oppure un vettore più lungo ma con un valore
unico ripetuto).}

\item{Nh}{vettore con la stima degli effettivi dello strato (può essere
un vettore di lunghezza uno oppure un vettore più lungo ma con un valore
unico ripetuto). In pratica è la somma dei pesi di campionamento dello
strato.}

\item{strata}{vettore che contiene le informazioni relative agli strati}
}
\description{
Estimated variance of a total estimation, with strata (vector version)
}
\examples{
tmp <- n_strata(d, weights = "w", strata = "s")
tmp$s <- 1L:3L
d <- merge(d, tmp, by = "s", all.x = TRUE)

# Adesso d ha le colonne mh e Nh
head(d)

k <- with(d, xcat == "cat2")
vhat3(d$w[k], d$mh[k], d$Nh[k], d$s[k])

}
