% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhat_mean.R
\name{vhat_mean_strata}
\alias{vhat_mean_strata}
\title{Estimate the variance of a mean estimate (of a strata)}
\usage{
vhat_mean_strata(y, weights, nPSU = NULL, na.rm = TRUE, fpc = TRUE)
}
\arguments{
\item{y}{numeric vector}

\item{weights}{numeric vector of weights}

\item{nPSU}{number of observations of the sample (default: NULL)}

\item{na.rm}{a logical value indicating whether NA values in y should be 
stripped before the computation proceeds. Default: TRUE}

\item{fpc}{finite population correction. Default: TRUE}
}
\description{
Estimate the variance of a mean estimate (of a strata)
}
\examples{
vhat_mean_strata(d$w, d$w, nPSU = nrow(d))

}
