% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{vhat_strata2}
\alias{vhat_strata2}
\title{Stima della varianza della stima di un totale di uno strato dei dati}
\usage{
vhat_strata2(data, weights, mh, Nh)
}
\arguments{
\item{data}{un data frame con il sottoinsieme dei dati per cui si vuole
stimare la varianza della stima di un totale. Da notare la differenza
rispetto all'argomento data della funzione \code{\link{vhat_strata}},
dove data rappresenta tutto il dataset, mentre qui rappresenta solo
il sottoinsieme dei dati la funzione indicatrice vale uno.}

\item{weights}{stringa con il nome della colonna dei dati che contiene i
pesi individuali}

\item{mh}{numero di osservazioni dello strato (vettore numerico di 
lunghezza uno).}

\item{Nh}{stima degli effettivi dello strato (vettore numerico di 
lunghezza uno).
In pratica è la somma dei pesi di campionamento dello strato.}
}
\description{
Stima della varianza della stima di un totale di uno strato dei dati
}
\examples{
tmp <- n_strata(d, weights = "w", strata = "s")
tmp
tmp$s <- 1L:3L

vhat_strata2(d[d$s == 2 & d$xcat == "cat2", ],
  weights = "w", mh = tmp$mh[2], Nh = tmp$Nh[2]
)

}
