% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{vhat_strata3}
\alias{vhat_strata3}
\title{Stima della varianza della stima del totale di un vettore x (uno strato)}
\usage{
vhat_strata3(x, mh, Nh)
}
\arguments{
\item{x}{un vettore numerico con gli elementi di uno strato (di solito i 
pesi di campionamento)}

\item{mh}{numero di osservazioni dello strato (vettore numerico di lunghezza uno).}

\item{Nh}{stima degli effettivi dello strato (vettore numerico di lunghezza uno).
In pratica è la somma dei pesi di campionamento dello strato.}
}
\description{
Stima della varianza della stima del totale di un vettore x (uno strato)
}
\examples{
tmp <- n_strata(d, weights = "w", strata = "s")
tmp
tmp$s <- 1L:3L
vhat_strata3(d$w[d$s == 2 & d$xcat == "cat2"], mh = tmp$mh[2], Nh = tmp$Nh[2])

}
