% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{vhat_strata4}
\alias{vhat_strata4}
\title{Stima della varianza della stima di un totale di uno strato dei dati}
\usage{
vhat_strata4(data, weights, mh, Nh)
}
\arguments{
\item{data}{un data frame con il sottoinsieme dei dati per cui si vuole 
stimare la varianza della stima di un totale. Da notare la differenza
rispetto all'argomento data della funzione \code{\link{vhat_strata}},
dove data rappresenta tutto il dataset, mentre qui rappresenta solo
il sottoinsieme dei dati la funzione indicatrice vale uno.}

\item{weights}{stringa con il nome della colonna dei dati che contiene i 
pesi individuali}

\item{mh}{nome della colonna dei dati che contiene il numero di osservazioni 
dello strato. Deve essere un vettore numerico con lo stesso valore ripetuto
il numero di righe dei dati.}

\item{Nh}{nome della colonna dei dati che contiene la stima degli effettivi 
dello strato. Come mh, deve essere lo stesso valore ripetuto n volte, dove
n è il numero di righe dei dati.}
}
\description{
Come vhat_strata2 ma mh e Nh sono due colonne dei dati di partenza.
}
\examples{
tmp <- n_strata(d, weights = "w", strata = "s")
tmp$s <- 1L:3L
d <- merge(d, tmp, by = "s", all.x = TRUE)

# Adesso d ha le colonne mh e Nh
vhat_strata4(d[d$s == 2 & d$xcat == "cat2", ], weights = "w", mh = "mh", Nh = "Nh")

}
