% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_strata.R
\name{vhat_strata5}
\alias{vhat_strata5}
\title{Stima della varianza della stima del totale di un vettore x (uno strato)}
\usage{
vhat_strata5(x, mh, Nh)
}
\arguments{
\item{x}{un vettore numerico con gli elementi di uno strato (di solito i 
pesi di campionamento)}

\item{mh}{vettore con il numero di osservazioni dello strato (può essere
un vettore di lunghezza uno oppure un vettore più lungo ma con un valore
unico ripetuto).}

\item{Nh}{vettore con la stima degli effettivi dello strato (può essere
un vettore di lunghezza uno oppure un vettore più lungo ma con un valore
unico ripetuto). In pratica è la somma dei pesi di campionamento dello
strato.}
}
\description{
come \code{\link{vhat_strata3}} ma con nh e Nh che possono anche essere 
vettori della stessa lunghezza di x (ma devono avere un valore unico...).
}
\examples{
tmp <- n_strata(d, weights = "w", strata = "s")
tmp$s <- 1L:3L
d <- merge(d, tmp, by = "s", all.x = TRUE)

# Adesso d ha le colonne mh e Nh
head(d)

k <- with(d, s == 2 & xcat == "cat2")
vhat_strata5(d$w[k], d$mh[k], d$Nh[k])

}
