% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhatbfs-package.R
\docType{package}
\name{vhatbfs-package}
\alias{vhatbfs}
\alias{vhatbfs-package}
\title{vhatbfs: Implementation of Confidence Intervals for Swiss Structural Survey Data}
\description{
Implementation of the FSO (Swiss Federal Statistical Office) formulas for estimating confidence intervals with Swiss Structrual Survey data. This mainly involves estimating confidence intervals of totals (estimating the numerosity of a population).
}
\keyword{internal}
