% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhatbfs.R
\name{vhatbfs}
\alias{vhatbfs}
\title{Implementazione degli Intervalli di Confidenza per i Dati RS.}
\description{
Implementazione delle formule dell'Ust per la stima degli intervalli di confidenza con i
dati RS. Si tratta principalmente di stimare degli intervalli di confidenza di effettivi
(stima della numerosità di una popolazione).
}
